+++ test/dummy/app/helpers/application_helper.rb
@@ -1,13 +1,4 @@
+# frozen_string_literal: true
+
 module ApplicationHelper
-  def presskits_extra_nav
-    render(
-      FlatPack::Button::Component.new(
-        text: "Sign out",
-        style: :ghost,
-        size: :md,
-        href: main_app.destroy_user_session_path,
-        data: { turbo_method: :delete }
-      )
-    )
-  end
 end