+++ test/dummy/config/initializers/recording_studio_presskits.rb
@@ -4,6 +4,7 @@ RecordingStudioPresskits.configure do |config|
   config.parent_root_type = "Workspace"
   config.authentication_method = :authenticate_user!
   config.current_actor_method = :current_user
+  config.excluded_picker_types = ["FakeBlock"]
 end
 
 RecordingStudioPresskits.register_section_component("FakeBlock", "FakeBlock::Component")