+++ test/dummy/test/test_helper.rb
@@ -17,8 +17,26 @@ module RoundedDefaultLayoutAssertions
   def assert_page_nav_close(href: "/recording_studio_presskits/press_kits")
     assert_select ".flat-pack-page-nav [data-flat-pack--icon-name-value='x-mark']", count: 1
     assert_select "a[href='#{href}'][aria-label='Close']", count: 1
-    assert_includes response.body, "Sign out"
     refute_includes response.body, "Dummy host"
+  end
+
+  def assert_access_slot_only
+    assert_includes response.body, "+ Access"
+    assert_includes response.body, "/accesses"
+    refute_includes response.body, "Sign out"
+    refute_includes response.body, ">Sign in<"
+    refute_select ".flat-pack-page-nav a[href*='recording_studio_root_switchable']"
+    refute_includes response.body, 'data-controller="recording-studio-root-switchable'
+  end
+
+  def assert_public_chrome_only
+    assert_select ".flat-pack-page-nav [data-flat-pack--icon-name-value='chevron-left']", count: 1
+    assert_select ".flat-pack-page-nav [data-flat-pack--icon-name-value='x-mark']", count: 1
+    refute_includes response.body, "Sign in"
+    refute_includes response.body, "Sign out"
+    refute_includes response.body, "+ Access"
+    refute_select ".flat-pack-page-nav a[href*='accesses']"
+    refute_select ".flat-pack-page-nav a[href*='recording_studio_root_switchable']"
   end
 end